+++ dataset/distribution/buyer_repetition.py
@@ -1,17 +1,15 @@
 """
 The most common buyer is identified in 1% to 50% of compiled releases.
 
-Failure indicates issues in buyer identification or buyer over-representation. Buyers are identified by
-``buyer.identifier.scheme`` and ``buyer.identifier.id``.
+Failure indicates issues in buyer identification or buyer over-representation. Buyers are identified by ``buyer.id``.
 
-The test is skipped if the ``buyer.identifier.scheme`` and ``buyer.identifier.id`` fields are both present in fewer
-than 1,000 compiled releases.
+The test is skipped if the ``buyer.id`` field is present in fewer than 1,000 compiled releases.
 """
 
 from pelican.util.checks import ReservoirSampler, get_empty_result_dataset
 from pelican.util.getter import deep_get
 
-version = 1.0
+version = 2.0
 min_items = 1000
 sample_size = 20
 
@@ -20,14 +18,13 @@ def add_item(scope, item, item_id):
     if not scope:
         scope = {"buyers": {}, "total_ocid_count": 0}
 
-    scheme = deep_get(item, "buyer.identifier.scheme")
-    ident = deep_get(item, "buyer.identifier.id")
-    if scheme is None or ident is None:
+    buyer_id = deep_get(item, "buyer.id")
+    if buyer_id is None:
         return scope
 
-    identifier = (str(scheme), str(ident))
-    scope["buyers"].setdefault(identifier, ReservoirSampler(sample_size))
-    scope["buyers"][identifier].process({"item_id": item_id, "ocid": item["ocid"]})
+    buyer_id = str(buyer_id)
+    scope["buyers"].setdefault(buyer_id, ReservoirSampler(sample_size))
+    scope["buyers"][buyer_id].process({"item_id": item_id, "ocid": item["ocid"]})
     scope["total_ocid_count"] += 1
 
     return scope
@@ -45,8 +42,8 @@ def get_result(scope):
         result["meta"] = {"reason": f"fewer than {min_items} occurrences of necessary fields"}
         return result
 
-    biggest_buyer_scheme, biggest_buyer_id = max(scope["buyers"], key=lambda key: scope["buyers"][key].index)
-    biggest_buyer = scope["buyers"][(biggest_buyer_scheme, biggest_buyer_id)]
+    biggest_buyer_id = max(scope["buyers"], key=lambda key: scope["buyers"][key].index)
+    biggest_buyer = scope["buyers"][biggest_buyer_id]
     biggest_buyer_share = biggest_buyer.index / total_ocid_count
     passed = 0.01 < biggest_buyer_share < 0.5
 
@@ -57,7 +54,7 @@ def get_result(scope):
         "ocid_count": biggest_buyer.index,
         "ocid_share": biggest_buyer_share,
         "examples": biggest_buyer.sample,
-        "specifics": {"buyer.identifier.scheme": biggest_buyer_scheme, "buyer.identifier.id": biggest_buyer_id},
+        "specifics": {"buyer.id": biggest_buyer_id},
     }
 
     return result